# api/frontend/openapi.yaml
openapi: 3.1.0
info:
  title: rtcv-ish Frontend API
  version: 0.1.0
  description: |
    REST + JSON between the browser and the rtcv-ish core, plus one
    Server-Sent Events stream (`GET /api/events`). See
    `design/frontend-api.md` for the rules.

    Errors are `Error` objects. `code` is one of:

    | code | HTTP | meaning |
    |---|---|---|
    | `INVALID_ARGUMENT` | 400 | Bad request body, parameter or value (also emulator `INVALID_ARGUMENT`). |
    | `OUT_OF_RANGE` | 400 | Address/size outside a domain (emulator `OUT_OF_RANGE`). |
    | `PERMISSION_DENIED` | 403 | The core cannot read a host path. |
    | `NOT_FOUND` | 404 | Unknown domain, stash key, slot content, list, ROM path, bundled emulator. |
    | `NO_ROM` | 409 | The emulator has no ROM loaded. |
    | `BUSY` | 409 | The emulator is busy, or another core operation is running and did not finish within 5 s (see `Status.busy`). |
    | `UNSUPPORTED` | 409 | The emulator lacks the capability. |
    | `ALREADY_CONNECTED` | 409 | Connect/launch while connected. |
    | `NO_BACKUP` | 409 | Toggle or protection "back" with nothing to restore. |
    | `CONNECT_FAILED` | 502 | Could not reach or handshake with the emulator. |
    | `FAILED` | 502 | The emulator reported a failure. |
    | `EMULATOR_DISCONNECTED` | 503 | The operation needs an emulator and none is connected. |
    | `EMULATOR_UNRESPONSIVE` | 503 | The emulator stopped answering (`Status.unresponsive`); disconnect, quit or wait for it to recover. |
...
servers:
...
paths:
  /status:
    get:
      operationId: getStatus
      tags: [status]
      responses:
        "200":
          description: Core and emulator status.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Status" }
        default: { $ref: "#/components/responses/Error" }
  /emulators:
    get:
      operationId: listEmulators
      tags: [emulator]
      description: Bundled emulators the core can launch.
      responses:
        "200":
          description: Bundled emulators.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/BundledEmulator" }
        default: { $ref: "#/components/responses/Error" }
  /emulator/connect:
    post:
      operationId: connectEmulator
      tags: [emulator]
      description: Connect to a running emulator. 409 `ALREADY_CONNECTED` when connected, 502 `CONNECT_FAILED` when unreachable.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ConnectRequest" }
      responses:
        "200":
          description: Connected.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Status" }
        default: { $ref: "#/components/responses/Error" }
  /emulator/launch:
    post:
      operationId: launchEmulator
      tags: [emulator]
      description: |
        Start a bundled emulator on a free port, connect to it and, when
        `rom` is set, load the ROM. The emulator process is owned by the
        core: it is quit when the core shuts down.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LaunchRequest" }
      responses:
        "200":
          description: Launched and connected.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Status" }
        default: { $ref: "#/components/responses/Error" }
  /emulator/disconnect:
    post:
      operationId: disconnectEmulator
      tags: [emulator]
      description: Drop the connection; the emulator keeps running. Idempotent.
      responses:
        "200":
          description: Disconnected.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Status" }
        default: { $ref: "#/components/responses/Error" }
  /emulator/rom:
    post:
      operationId: loadRom
      tags: [emulator]
      description: |
        Load a ROM by path on the emulator host. Clears scheduled units,
        refreshes the domain list and auto-selects the non-hidden domains.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PathRequest" }
      responses:
        "200":
          description: ROM loaded.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GameStatus" }
        default: { $ref: "#/components/responses/Error" }
  /emulator/close-rom:
    post:
      operationId: closeRom
      tags: [emulator]
      responses:
        "200":
          description: ROM closed.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GameStatus" }
        default: { $ref: "#/components/responses/Error" }
  /emulator/pause:
    post:
      operationId: pauseEmulator
      tags: [emulator]
      responses:
        "200":
          description: Paused.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GameStatus" }
        default: { $ref: "#/components/responses/Error" }
  /emulator/resume:
    post:
      operationId: resumeEmulator
      tags: [emulator]
      responses:
        "200":
          description: Running.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GameStatus" }
        default: { $ref: "#/components/responses/Error" }
  /emulator/reset:
    post:
      operationId: resetEmulator
      tags: [emulator]
      responses:
        "200":
          description: Reset.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GameStatus" }
        default: { $ref: "#/components/responses/Error" }
  /emulator/step:
    post:
      operationId: stepEmulator
      tags: [emulator]
      description: Pause (if running) and emulate `frames` frames.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/StepRequest" }
      responses:
        "200":
          description: Stepped; the emulator is paused.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GameStatus" }
        default: { $ref: "#/components/responses/Error" }
  /emulator/quit:
    post:
      operationId: quitEmulator
      tags: [emulator]
      description: Ask the emulator to exit. The core is disconnected afterwards.
      responses:
        "204": { description: Quit sent. }
        default: { $ref: "#/components/responses/Error" }
  /emulator/screenshot:
    get:
      operationId: getScreenshot
      tags: [emulator]
      responses:
        "200":
          description: PNG of all screens stacked vertically.
          content:
            image/png:
              schema: { type: string, format: binary }
        default: { $ref: "#/components/responses/Error" }
  /domains:
    get:
      operationId: listDomains
      tags: [memory]
      responses:
        "200":
          description: Memory domains of the loaded game (empty without a ROM).
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Domain" }
        default: { $ref: "#/components/responses/Error" }
  /domains/selected:
    put:
      operationId: setSelectedDomains
      tags: [memory]
      description: Replace the selection used by generation. Unknown names are 400 `INVALID_ARGUMENT`.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/NamesRequest" }
      responses:
        "200":
          description: Updated domain list.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Domain" }
        default: { $ref: "#/components/responses/Error" }
  /domains/auto-select:
    post:
      operationId: autoSelectDomains
      tags: [memory]
      description: Select all non-hidden domains.
      responses:
        "200":
          description: Updated domain list.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Domain" }
        default: { $ref: "#/components/responses/Error" }
  /memory/{domain}:
    parameters:
      - { $ref: "#/components/parameters/DomainName" }
    get:
      operationId: readMemory
      tags: [memory]
      parameters:
        - name: address
          in: query
          required: true
          schema: { type: integer, format: int64, minimum: 0 }
        - name: size
          in: query
          required: true
          schema: { type: integer, format: int64, minimum: 1, maximum: 65536 }
      responses:
        "200":
          description: Memory contents.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MemoryChunk" }
        default: { $ref: "#/components/responses/Error" }
    put:
      operationId: writeMemory
      tags: [memory]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MemoryWrite" }
      responses:
        "204": { description: Written. }
        default: { $ref: "#/components/responses/Error" }
  /memory/{domain}/words:
    parameters:
      - { $ref: "#/components/parameters/DomainName" }
    get:
      operationId: readMemoryWords
      tags: [memory]
      description: |
        Raw 16-bit words for the memory visualizer. Returns every
        `stride`-th little-endian word of `[address, address+size)`, in
        memory order and byte order: word i is the 2 bytes at
        `address + 2*stride*i`. The response has
        `2*ceil(floor(size/2)/stride)` bytes; a trailing odd byte is
        ignored. The range must lie inside the domain (400
        `OUT_OF_RANGE`); there is no 64 KiB cap. The core reads the
        emulator in batches of at most 1 MiB.
      parameters:
        - name: address
          in: query
          required: false
          schema: { type: integer, format: int64, minimum: 0, default: 0 }
        - name: size
          in: query
          required: true
          schema: { type: integer, format: int64, minimum: 2 }
        - name: stride
          in: query
          required: false
          schema: { type: integer, format: int64, minimum: 1, maximum: 65536, default: 1 }
      responses:
        "200":
          description: Sampled words, little-endian, 2 bytes each.
          content:
            application/octet-stream:
              schema: { type: string, format: binary }
        default: { $ref: "#/components/responses/Error" }
  /settings:
    get:
      operationId: getSettings
      tags: [settings]
      responses:
        "200":
          description: The whole settings object.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Settings" }
        default: { $ref: "#/components/responses/Error" }
    patch:
      operationId: patchSettings
      tags: [settings]
      description: |
        Deep merge: only the given fields change, nested objects are merged
        field by field. The result is validated as a whole (400
        `INVALID_ARGUMENT`, nothing changes) and persisted to
        `settings.json`. `autoCorrupt` is not persisted; it is false at
        start.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SettingsPatch" }
      responses:
        "200":
          description: The whole settings object after the update.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Settings" }
        default: { $ref: "#/components/responses/Error" }
  /blast:
    post:
      operationId: manualBlast
      tags: [blast]
      description: Generate a layer with the current settings on the selected domains and apply it (no backup). 409 `NO_ROM` without a ROM, 400 `INVALID_ARGUMENT` when no domain is selected.
      responses:
        "200":
          description: The generated layer.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Layer" }
        default: { $ref: "#/components/responses/Error" }
  /blast/apply:
    post:
      operationId: applyLayer
      tags: [blast]
      description: Apply a layer to the running game. With `backup`, the current bytes at every unit are stored for `/blast/toggle`.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ApplyRequest" }
      responses:
        "204": { description: Applied. }
        default: { $ref: "#/components/responses/Error" }
...
  /savestates:
    get:
      operationId: listSavestates
      tags: [savestates]
      responses:
        "200":
          description: All savestate slots, numbered 1..50, in order.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/SavestateSlot" }
        default: { $ref: "#/components/responses/Error" }
...
components:
...
  responses:
    Error:
      description: Error.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...